+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 * Add `AutoAccepted` status to `Dispute` 
 * Add `DisputeAutoAccepted` webhook to `WebhookNotification`
 * Add `ExcessiveRetry` to `GatewayRejectionReason`
+* Add `SEPADirectDebitAccount` payment method
+* Add `sepa_direct_debit_return_code` to `Transaction`
+* Add `sepa_direct_debit_account_details` to `Transaction`
+* Add `sepa_direct_debit_account_nonce_details` to `PaymentMethodNonceDetails`
+* Add `SEPA_DIRECT_DEBIT_ACCOUNT` to `PaymentInstrumentType`
+* Add `sepa_debit_paypal_v2_order_id` to `TransactionSearch`
+* Add `sepa_direct_debit_accounts` to `Customer`
+* Add SEPA Direct Debit specific error codes
 
 ## 4.9.0
 * Add `cardholder_name` attribute reader to `ApplePayCard` (Thanks @AxelTheGerman)